requirements.txt
flask
flask-cors
python-dateutil
requests
plyer
